requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
commonmark==0.9.1
Deprecated==1.2.18
Flask==2.3.2
Flask-Caching==1.10.1
Flask-Cors==5.0.0
Flask-Limiter==2.4.2
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-sqlalchemy==1.4.0
mysql-connector-python==8.0.27
packaging==24.2
protobuf==5.29.3
Pygments==2.19.1
PyJWT==2.4.0
python-dotenv==0.19.2
rich==12.6.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
